Moved getBounds to Groupfigure class, as it is general for all figures. Furhtermore cleaned up the method using early return
