+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/figures/SVGGroupFigure.java
@@ -8,17 +8,18 @@
 package org.jhotdraw.samples.svg.figures;
 
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.draw.AttributeKey;
 import org.jhotdraw.draw.figure.GroupFigure;
-import org.jhotdraw.draw.figure.Figure;
-import java.awt.*;
-import java.awt.geom.*;
-import java.util.*;
-import org.jhotdraw.draw.*;
-import static org.jhotdraw.draw.AttributeKeys.TRANSFORM;
 import org.jhotdraw.draw.figure.SVGOpacityStrategy;
 import org.jhotdraw.draw.handle.Handle;
 import org.jhotdraw.draw.handle.TransformHandleKit;
 import org.jhotdraw.samples.svg.SVGAttributeKeys;
+
+import java.awt.*;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.Map;
+
 import static org.jhotdraw.samples.svg.SVGAttributeKeys.*;
 
 /**
@@ -73,31 +74,9 @@ public class SVGGroupFigure extends GroupFigure implements SVGFigure {
     public void draw(Graphics2D g) {
         double opacity = get(OPACITY);
         SVGOpacityStrategy ops = new SVGOpacityStrategy(this);
-        if (ops.execute(g, opacity)){
+        if (ops.execute(g, opacity)) {
             super.draw(g);
         }
-    }
-
-    @Override
-    public Rectangle2D.Double getBounds() {
-        if (cachedBounds == null) {
-            if (getChildCount() == 0) {
-                cachedBounds = new Rectangle2D.Double();
-            } else {
-                for (Figure f : children) {
-                    Rectangle2D.Double bounds = f.getBounds();
-                    if (f.get(TRANSFORM) != null) {
-                        bounds.setRect(f.get(TRANSFORM).createTransformedShape(bounds).getBounds2D());
-                    }
-                    if (cachedBounds == null) {
-                        cachedBounds = bounds;
-                    } else {
-                        cachedBounds.add(bounds);
-                    }
-                }
-            }
-        }
-        return (Rectangle2D.Double) cachedBounds.clone();
     }
 
     @Override
